Resizable panel with reasonable default width
Some of my collaborators have long last names which makes a dynamic width
annoying. Because of this (and another reason, see below) I decided to
implement a static-width panel that's resizable.

This commit adds a resize-handle and some style tweaks to make things look good
at varying widths.

The other reason a static-width panel is better is because when blaming large
files in repositories with lots of history there is a noticeable delay before
the panel appears. It would feel much better if it displayed instantly, but
this is problematic in the current implementation because the width is unknown
until blame returns.

With a static panel width, the panel can react immediately to the keypress and
display a throbber to indicate blame is running. Then when blame completes, the
panel can update without resizing.

I haven't actually implemented instant display yet but I intend to.

Test Plan:

 * Opened blame, saw long names had ellipsis
 * Resized panel with mouse
 * Scrolled down and up, resize-handle still usable
 * Edited some lines and made sure 'no commit' lines look OK