Rewrite gutter views in React
This will make it much easier to update the view after it becomes visible. This
commit is just a rewrite though and I haven't added any functionality yet.

This commit is dependent on atom/atom#3102 which allows me to workaround
facebook/react#1939. I might vendorize a hacked version of React if those bugs
take a while to be fixed.

Also I noticed one regression - clicking in the editor seems to have a slight
delay when the gutter is open. I'm not sure what's causing this but will try
to fix.

Test plan:

I tested every piece of functionality I could think of, including:

  * Clicking on hash links
  * Scrolling
  * Opening when already scrolled
  * Adjusting width
  * Opening on a file that's not checked in (error still appears)