Instant blame display, remove key throttling
This make the blame window appear instantly when the key is pressed. It feels
nicer like this because there's instant feedback that you hit the correct key.
Same goes for closing the window. Throttling/debouncing is completely removed.

Test Plan:
Open a new Atom window and press `ctrl-b`. Now the window opens and closes
instantly. The data loads in the background if necessary.

I also held down `ctrl-b` with a really small key-repeat delay and it seems to
work pretty robustly.